feat(libwrapper): modify default effects tray
fixes #18
fixes #19
fixes #16

might also actually help with #14
