+++ lang/en.json
@@ -3,7 +3,7 @@
     "AllowTargetSettingName": "Transfer to Target",
     "AllowTargetSettingHint": "Allow users to transfer effects to targets they do not own.",
     "DamageDefaultSettingName": "Use Default Damage Tray",
-    "DamageDefaultSettingHint": "Retain default system behavior (only the GM can see damage trays).",
+    "DamageDefaultSettingHint": "Retain default system behavior for damage trays (DM only trays).",
     "DamageTargetSettingName": "Damage Target",
     "DamageTargetSettingHint": "Allow users to damage targets that they do not own with the damage tray.",
     "DeleteInsteadSettingName": "Delete Instead of Refresh",
@@ -23,9 +23,7 @@
       "NoActiveGMEffect": "There is no active GM. An active GM is required to create effects on targets you do not own.",
       "NoTarget": "You don't have a target."
     },
-    "ScrollOnExpandSettingName": "Scroll on Expand",
-    "ScrollOnExpandSettingHint": "Scroll chat to bottom when expanding a tray that is at the bottom.",
     "SystemDefaultSettingName": "Use Default Effects Tray",
-    "SystemDefaultSettingHint": "Retain default system behavior (if you only want the settings below this one)."
+    "SystemDefaultSettingHint": "Retain default system behavior for effect application trays (DM only trays outside of transfer effects)."
   }
 }
