+++ scripts/setup.mjs
@@ -21,10 +21,7 @@
 import { ModuleSettings } from "./settings.mjs";
 import { EffectiveTray } from "./effective-tray.mjs";
 import { EffectiveSocket } from "./effective-socket.mjs";
-import EffectiveEAE from "./effect-application.mjs";
 import { API } from "./api.mjs";
-
-window.customElements.define("effective-effect-application", EffectiveEAE);
 
 Hooks.once("init", EffectiveSocket.init);
 Hooks.once("init", ModuleSettings.init);
